+++ app/(tabs)/home/index.tsx
@@ -1,6 +1,6 @@
 import { SafeAreaView, ScrollView } from "react-native";
 import SearchBar from "@/components/common/SearchBar";
-import HomeSection from "@/components/ui/home/HomeSection";
+import HomeSection from "@/components/ui/home/index/Section";
 
 export default function HomeScreen() {
   return (
