+++ app/(tabs)/home/place-trending.tsx
@@ -1,10 +1,10 @@
 import { SafeAreaView, ScrollView, View } from "react-native";
 import { usePlaceListData } from "@/hooks/query/placeData";
-import PlaceCard from "@/components/ui/home/PlaceCard";
+import PlaceCard from "@/components/ui/home/list/PlaceCard";
 
 export default function PlaceTrendingScreen() {
   const { data: placeList } = usePlaceListData("place-trending");
-  
+
   return (
     <SafeAreaView tw="flex-1 mb-20">
       <ScrollView tw="pt-2">
